Fix: 로그인 API 수정
response에 이름 추가

#17
