+++ src/main/java/com/kimdabang/kdbserver/auth/application/AuthService.java
@@ -1,6 +1,10 @@
 package com.kimdabang.kdbserver.auth.application;
 
-import com.kimdabang.kdbserver.auth.dto.*;
+import com.kimdabang.kdbserver.auth.dto.in.SignInRequestDto;
+import com.kimdabang.kdbserver.auth.dto.in.SignUpRequestDto;
+import com.kimdabang.kdbserver.auth.dto.in.TestTokenRequestDto;
+import com.kimdabang.kdbserver.auth.dto.out.SignInResponseDto;
+import com.kimdabang.kdbserver.auth.dto.out.TestTokenResponseDto;
 
 public interface AuthService {
