+++ src/main/java/com/kimdabang/kdbserver/auth/dto/in/SignInRequestDto.java
@@ -1,4 +1,4 @@
-package com.kimdabang.kdbserver.auth.dto;
+package com.kimdabang.kdbserver.auth.dto.in;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,6 +9,9 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class TestTokenRequestDto {
-    private String accessToken;
+public class SignInRequestDto {
+
+    private String loginId;
+    private String password;
+
 }
